Move test and add TODO
